requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9

langchain>=0.2.0
langchain-community>=0.2.0
langchain-openai>=0.1.0
langchain-pinecone>=0.1.0
langchain-text-splitters>=0.2.0

openai>=1.30.0
pinecone-client>=4.1.0
pypdf>=4.2.0

streamlit>=1.35.0
requests>=2.31.0
python-dotenv>=1.0.0
pydantic>=2.7.0
